move number-input directory
